Renamed and simplified variant constructors
Removed the `Leaf` constructor for constants. A constant can be
represented as a nullary function application just as well. The previous
invariant that a function application has a non-empty list of arguments
does not hold any more, but it was never used anyways.

Renamed `Node` to `App` and `Annot` to `Attr` in the variant type of
terms. This is the same constructor as for flat terms, but that is fine
in recent versions of OCaml.