Generate range constraints as assertions
Instead of changing the type of the state variable to a stricter type, add an assertion if the right-hand side of an equation has been inferred to a stricter type than the left-hand side.

If we have an equation

    x = if in then 0 else 1;

for the variable

    x: int;

then the type of the right-hand side is inferred to ```subrange [0; 1] of int```. WE make use of that information and generate the assertion

    0 <= x and x <= 1